+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.1] - 2025-01-24
+
 ### Changed
 
 - constraint-set.hpp : remove template-id for destructor
@@ -155,7 +157,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * This is the first release of `proxsuite-nlp`. This library is a joint effort between LAAS-CNRS and INRIA, and will be maintained and expanded in the future. Please provide constructive feedback and contribute!
 
-[Unreleased]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.10.1...HEAD
+[0.10.1]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/Simple-Robotics/proxsuite-nlp/compare/v0.7.1...v0.8.0